+++ pkgs/test/test/runner/configuration/platform_test.dart
@@ -41,7 +41,7 @@ void main() {
       expect(
           test.stdout,
           containsInOrder(
-              ['-1: [Chrome] test [E]', '+1 -1: Some tests failed.']));
+              ['-1: [Chrome, Dart2Js] test [E]', '+1 -1: Some tests failed.']));
       await test.shouldExit(1);
     }, tags: ['chrome']);
 
@@ -70,8 +70,8 @@ void main() {
       expect(
           test.stdout,
           containsInOrder([
-            '-1: [Chrome] test [E]',
-            '-2: [VM] test [E]',
+            '-1: [Chrome, Dart2Js] test [E]',
+            '-2: [VM, Kernel] test [E]',
             '-2: Some tests failed.'
           ]));
       await test.shouldExit(1);