+++ pkgs/test_api/test/utils.dart
@@ -21,7 +21,8 @@ import 'package:test_core/src/runner/runner_suite.dart';
 import 'package:test_core/src/runner/suite.dart';
 
 /// A dummy suite platform to use for testing suites.
-final suitePlatform = SuitePlatform(Runtime.vm);
+final suitePlatform =
+    SuitePlatform(Runtime.vm, compiler: Runtime.vm.defaultCompiler);
 
 // The last state change detected via [expectStates].
 State? lastState;